requirements.txt
Django==5.2.1
django-extensions==4.1

psycopg2
pyyml
requests
phonenumbers

gpxpy==1.6.2
overpy==0.7
numpy
pandas

pycodestyle

redis
celery==5.5.2

geojson==3.2.0
djangorestframework==3.16.0
django-cors-headers==4.7.0

social-auth-app-django~=5.4.3

geopy
smopy

gunicorn
